requirements.txt
django>=3.2.0,<=3.2.99
mssql-django
django-cors-headers
djangorestframework>=3.12.0,<=3.12.99
djangorestframework-queryfields>=1.0.0
djangorestframework-filters==1.0.0.dev0
git+https://github.com/Innovate-Inc/django-agol-oauth2.git@master#egg=django-agol-oauth2==0.0.3
django-rest-framework-social-oauth2==1.1.0
python-dotenv
wfastcgi==3.0.0
pywinpty==0.5.1
geojson
arcgis==1.8.1
